+++ skrub/_apply_to_cols.py
@@ -3,18 +3,19 @@ ApplyToCols selects the correct transformer between ApplyToEachCol and ApplyToSu
 based on the type of the transformer passed to it.
 """
 
-from sklearn.base import BaseEstimator, TransformerMixin, check_is_fitted
+from sklearn.base import TransformerMixin, check_is_fitted
 
 from . import selectors
 from ._apply_to_each_col import ApplyToEachCol
 from ._apply_to_sub_frame import ApplyToSubFrame
+from ._base import BaseTransformer
 from ._sklearn_compat import _VisualBlock
 from ._wrap_transformer import wrap_transformer
 
 _SELECT_ALL_COLUMNS = selectors.all()
 
 
-class ApplyToCols(TransformerMixin, BaseEstimator):
+class ApplyToCols(TransformerMixin, BaseTransformer):
     """
     Apply a transformer to selected columns in a dataframe.
 
@@ -291,20 +292,6 @@ class ApplyToCols(TransformerMixin, BaseEstimator):
     0 -10.0    0.0      -1.0      -1.0
     1  10.0  100.0       1.0       1.0
     """  # noqa: E501
-
-    _doc_link_module = "skrub"
-
-    # Defining this as a property because it inherits from _HTMLDocumentationLinkMixin,
-    # which also defines _doc_link_template as a property, and we want to be able
-    # to override it.
-    @property
-    def _doc_link_template(self):
-        return getattr(
-            self,
-            "__doc_link_template",
-            "https://skrub-data.org/stable/reference/generated/"
-            "{estimator_module}.{estimator_name}.html",
-        )
 
     def __init__(
         self,
