+++ skrub/_squashing_scaler.py
@@ -5,6 +5,7 @@ from sklearn.base import BaseEstimator, OneToOneFeatureMixin, TransformerMixin
 from sklearn.preprocessing import RobustScaler
 from sklearn.utils.validation import FLOAT_DTYPES, check_is_fitted
 
+from skrub._base import BaseTransformer
 from skrub._sklearn_compat import validate_data
 
 
@@ -82,7 +83,7 @@ class _MinMaxScaler(OneToOneFeatureMixin, TransformerMixin, BaseEstimator):
         return self.scale_ * (X - self.median_)
 
 
-class SquashingScaler(OneToOneFeatureMixin, TransformerMixin, BaseEstimator):
+class SquashingScaler(OneToOneFeatureMixin, TransformerMixin, BaseTransformer):
     r"""Perform robust centering and scaling followed by soft clipping.
 
     When features have large outliers, smooth clipping prevents the outliers from
@@ -335,17 +336,3 @@ class SquashingScaler(OneToOneFeatureMixin, TransformerMixin, BaseEstimator):
             X_tr = _set_zeros(X_tr, self.zero_cols_)
 
         return _soft_clip(X_tr, self.max_absolute_value, mask_inf)
-
-    _doc_link_module = "skrub"
-
-    # Defining this as a property because it inherits from _HTMLDocumentationLinkMixin,
-    # which also defines _doc_link_template as a property, and we want to be able
-    # to override it.
-    @property
-    def _doc_link_template(self):
-        return getattr(
-            self,
-            "__doc_link_template",
-            "https://skrub-data.org/stable/reference/generated/"
-            "{estimator_module}.{estimator_name}.html",
-        )
